+++ app/models/unidom/category/categorizing.rb
@@ -1,6 +1,6 @@
 # Categorizing 是分类（category）和被分类项（categorized）之间的关联关系。
 
-class Unidom::Category::Categorizing < ::ActiveRecord::Base
+class Unidom::Category::Categorizing < ActiveRecord::Base
 
   self.table_name = 'unidom_categorizings'
 
@@ -9,7 +9,11 @@ class Unidom::Category::Categorizing < ::ActiveRecord::Base
   belongs_to :category,    class_name:  'Unidom::Category::Category'
   belongs_to :categorized, polymorphic: true
 
-  scope :category_is,    ->(category)    { where category_id: (category.respond_to?(:id) ? category.id : category) }
-  scope :categorized_is, ->(categorized) { where categorized: categorized }
+  scope :category_is,    ->(category)    { where category_id: to_id(category) }
+  scope :categorized_is, ->(categorized) { where categorized: categorized     }
+
+  def self.categorize!(category, categorized, opened_at = Time.now)
+    self.categorized_is(categorized).category_is(category).valid_at.alive.first_or_create! elemental: true, opened_at: opened_at
+  end
 
 end