+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 2. Category model & migration
 3. Category Rollup model & migration
 4. Categorizing model & migration
+
+## v0.2
+1. Improved the Categorizing model to add the ::categorize! method
+2. Improved the Category model for the validation
+3. Improved the Ruby Gem Specification to depend on unidom-common v0.9